requirements.txt
# Core Federated Learning
flwr[simulation]>=1.18.0

# FedRAG (optional - for advanced features)
# fed-rag[qdrant,llamaindex]>=0.0.27

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# RAG Framework
llama-index>=0.10.0
llama-index-vector-stores-qdrant>=0.2.0
llama-index-embeddings-huggingface>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Vector Database
qdrant-client>=1.7.0

# LLM APIs
openai>=1.10.0
anthropic>=0.18.0

# Data Processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# NLP
spacy>=3.5.0
scispacy>=0.5.0

# Medical NLP (optional)
# en-core-sci-sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz

# Text Processing
nltk>=3.8.0
regex>=2023.0.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gradio>=4.0.0
streamlit>=1.30.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.0
redis>=5.0.0
sqlalchemy>=2.0.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.0

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
httpx>=0.24.0

# Docker & Deployment
gunicorn>=21.2.0
docker>=6.1.0
